requirements-test.txt
# テスト依存関係
pytest==7.4.3
pytest-mock==3.12.0
pytest-asyncio==0.21.1
pytest-cov==4.1.0
moto==4.2.14  # AWS サービスのモック
PyJWT==2.8.0  # JWT処理
requests-mock==1.11.0  # HTTPリクエストのモック